+++ packages/rocketh-signer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rocketh/signer
 
+## 0.15.0-testing.3
+
+### Patch Changes
+
+- fixes
+
 ## 0.15.0-testing.2
 
 ### Patch Changes