+++ packages/rocketh/CHANGELOG.md
@@ -1,5 +1,11 @@
 # rocketh
 
+## 0.15.0-testing.14
+
+### Patch Changes
+
+- fixes
+
 ## 0.15.0-testing.13
 
 ### Patch Changes